% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpa_summary_maker.R
\name{hpa_summary_maker}
\alias{hpa_summary_maker}
\title{hpa_summary_maker}
\usage{
hpa_summary_maker(hpa_dat)
}
\arguments{
\item{tissue_type}{A character string that determines which HPA data you want to download from the website. Has to be
"both" (default), "normal", or "cancer".}

\item{save_file}{A boolean determining if you want the HPA data downloaded permanently or temporarily. Default is TRUE,
meaning the file will be saved in the given "save_location", default being the current working directory.}

\item{save_location}{A character string indicating where you want the files to be saved if you are saving them.
If the file(s) already exists in that location, those will be loaded instead of re-downloading the files.}
}
\value{
A dataframe summarizing the amount of proteins tested to detected, used for the shiny app.
}
\description{
Used to generate a summary file used in the second tab of the Shiny app version of HPAStainR
}
\examples{
## Generate the summarized HPA file
hpa_summary <- hpa_summary_maker(hpa_dat)
}
