% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPA_summary_maker.R
\name{HPA_summary_maker}
\alias{HPA_summary_maker}
\title{HPA_summary_maker}
\usage{
HPA_summary_maker(hpa_dat)
}
\arguments{
\item{hpa_dat}{The dataframe of normal tissue data downloaded by HPA_data_downloader()}
}
\value{
A dataframe summarizing the amount of proteins tested to detected, used for the shiny app.
}
\description{
Used to generate a summary file used in the second tab of the Shiny app version of HPAStainR
}
\examples{
## Load in data from downloader
HPA_data <- HPA_data_downloader(save_file = FALSE)
## Generate the summarized HPA file
hpa_summary <- HPA_summary_maker(HPA_data$hpa_dat)
}
